+++ scripts/app/configs/paths.ts
