+++ scripts/app/exceptions/Exception.ts
