+++ scripts/app/exceptions/FileNotFoundException.ts
