+++ scripts/start.ts
@@ -24,8 +24,8 @@ import * as openBrowser from "react-dev-utils/openBrowser";
 import * as WebpackDevServerUtils from "react-dev-utils/WebpackDevServerUtils";
 import * as webpack from "webpack";
 import * as WebpackDevServer from "webpack-dev-server";
-import Application from "./Application";
-import WebpackDevServerConfg from "./WebpackDevServerConfg";
+import Application from "./app/Application";
+import WebpackDevServerConfig from "./app/WebpackDevServerConfig";
 
 /**
  * Get WebpackDevServerUtils functions
@@ -98,7 +98,7 @@ try {
         const urls = prepareUrls(server.protocol, server.host, port);
         const compiler = createCompiler(webpack, config, appName, urls, usingYarn);
         const proxyConfig = prepareProxy(proxySettings, publicPath);
-        const webpackDevServerConfg = new WebpackDevServerConfg(server, publicPath, srcPath);
+        const webpackDevServerConfg = new WebpackDevServerConfig(server, publicPath, srcPath);
         const serverConfig = webpackDevServerConfg.create(
             proxyConfig,
             urls.lanUrlForConfig,
